requirements.txt
Django==1.8.2
Fabric==1.10.1
bcrypt==1.1.1
django-appconf==1.0.1
django-compressor==1.5
django-cors-headers==1.1.0
django-filter==0.10.0
django-nose==1.4
djangorestframework==3.1
flake8==2.4.0
gunicorn==19.3
nose==1.3.6
psycopg2==2.6
